feat: add CloudWatch metric alarm ARN outputs
This commit adds new output variables to provide a map of FSx filesystem identifiers and their corresponding CloudWatch metric alarm ARNs. The changes include modifications to the main.tf and outputs.tf files in the examples/01_default_configuration and examples/02_custom_schedule directories, as well as the outputs.tf file in the root directory.